@@ -1,55 +1,43 @@
 {
-  "name": "frontend",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.0",
-    "@testing-library/user-event": "^13.2.1",
-    "bootstrap": "^5.1.1",
-    "formik": "^2.1.5",
-    "fp-ts": "^2.11.3",
-    "io-ts": "^2.2.16",
-    "io-ts-types": "^0.5.16",
-    "papaparse": "^5.3.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0-rc.0",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "yup": "^0.32.9"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "serveprod": "yarn build && http-server --port 8000 --proxy http://localhost:8000? ./build"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@types/node": "^16.9.6",
-    "@types/papaparse": "^5.2.6",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.3.0",
-    "@types/yup": "^0.29.13",
-    "http-server": "^13.0.2",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.4.3"
-  }
+	"name": "vite-project",
+	"private": true,
+	"version": "0.0.0",
+	"type": "module",
+	"scripts": {
+		"dev": "vite",
+		"build": "tsc && vite build",
+		"lint": "tsc --noEmit && eslint --fix src",
+		"preview": "vite preview"
+	},
+	"dependencies": {
+		"@popperjs/core": "^2.11.6",
+		"bootstrap": "^5.2.2",
+		"formik": "^2.2.9",
+		"fp-ts": "^2.13.1",
+		"io-ts": "^2.2.19",
+		"papaparse": "^5.3.2",
+		"react": "^18.2.0",
+		"react-bootstrap": "^2.5.0",
+		"react-dom": "^18.2.0",
+		"react-router-dom": "^6.4.2",
+		"yup": "^0.32.11"
+	},
+	"devDependencies": {
+		"@trivago/prettier-plugin-sort-imports": "^3.3.1",
+		"@types/papaparse": "^5.3.5",
+		"@types/react": "^18.0.17",
+		"@types/react-dom": "^18.0.6",
+		"@typescript-eslint/eslint-plugin": "^5.40.1",
+		"@typescript-eslint/parser": "^5.40.1",
+		"@vitejs/plugin-react": "^2.1.0",
+		"eslint": "^8.25.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-prettier": "^4.2.1",
+		"eslint-plugin-react": "^7.31.10",
+		"prettier": "^2.7.1",
+		"prettier-config-standard": "^5.0.0",
+		"sass": "^1.55.0",
+		"typescript": "^4.6.4",
+		"vite": "^3.1.0"
+	}
 }
