@@ -1,69 +1,67 @@
-import React, { useState } from "react";
-import { Button, Form, Modal } from "react-bootstrap";
-import requestsWrapper from "../../RequestsWrapper";
+import { useState } from 'react'
+import { Button, Form, Modal } from 'react-bootstrap'
+
+import requestsWrapper from '../../RequestsWrapper'
 
 interface AddUserModalProps {
-  show: boolean;
-  setShow(b: boolean): void;
-  courseName: string;
+	show: boolean
+	setShow(b: boolean): void
+	courseName: string
 }
 
 const AddUserModal = ({ show, setShow, courseName }: AddUserModalProps) => {
-  const [username, setUsername] = useState("");
-  const [message, setMessage] = useState<{
-    header: string;
-    message: string;
-  } | null>(null);
-  const handleClose = () => setShow(false);
-  return (
-    <>
-      <Modal show={show} onHide={() => {}}>
-        <Modal.Header>
-          <Modal.Title>Adding user to "{courseName}"</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          Enter username
-          <Form.Control
-            type="text"
-            value={username}
-            onChange={(event) => setUsername(event.target.value)}
-          ></Form.Control>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
-            Close
-          </Button>
-          <Button
-            variant="primary"
-            onClick={async () => {
-              const res = await requestsWrapper.addStudent(
-                username,
-                courseName
-              );
-              if (res._tag === "Right") {
-                setMessage({ header: "Success", message: res.right });
-              } else {
-                setMessage({ header: "Error", message: res.left.toString() });
-              }
-            }}
-          >
-            Add user
-          </Button>
-        </Modal.Footer>
-      </Modal>
-      <Modal show={message !== null} onHide={() => {}}>
-        <Modal.Header>
-          <Modal.Title>{message?.header}</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>{message?.message}</Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={() => setMessage(null)}>
-            Close
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </>
-  );
-};
+	const [username, setUsername] = useState('')
+	const [message, setMessage] = useState<{
+		header: string
+		message: string
+	} | null>(null)
+	const handleClose = () => setShow(false)
+	return (
+		<>
+			<Modal show={show}>
+				<Modal.Header>
+					<Modal.Title>Adding user to &quot;{courseName}&quot;</Modal.Title>
+				</Modal.Header>
+				<Modal.Body>
+					Enter username
+					<Form.Control
+						type="text"
+						value={username}
+						onChange={(event) =>
+							setUsername(event.target.value)
+						}></Form.Control>
+				</Modal.Body>
+				<Modal.Footer>
+					<Button variant="secondary" onClick={handleClose}>
+						Close
+					</Button>
+					<Button
+						variant="primary"
+						onClick={async () => {
+							const res = await requestsWrapper.addStudent(username, courseName)
+							if (res._tag === 'Right') {
+								setMessage({ header: 'Success', message: res.right })
+							} else {
+								setMessage({ header: 'Error', message: res.left.toString() })
+							}
+						}}>
+						Add user
+					</Button>
+				</Modal.Footer>
+			</Modal>
+			<Modal show={message !== null}>
+				<Modal.Header>
+					<Modal.Title>{message?.header}</Modal.Title>
+				</Modal.Header>
+				<Modal.Body>{message?.message}</Modal.Body>
+				<Modal.Footer>
+					<Button variant="secondary" onClick={() => setMessage(null)}>
+						Close
+					</Button>
+				</Modal.Footer>
+			</Modal>
+		</>
+	)
+}
 
-export default AddUserModal;
+export default AddUserModal
